split message result into own file for readability
