+++ test/rust-test-util/TestResultReportingClient/src/models/mod.rs
@@ -1,1 +1,2 @@
-pub mod test_result;
+pub mod message_result;
+pub mod test_result_dto;
